requirements.txt
flask==3.0.3
flask-cors==4.0.1
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.16.1
gunicorn==22.0.0
python-dotenv==1.0.1